# openapi.json
{"openapi": "3.1.0", "info": {"title": "3rd party", "summary": "메타데이터 기반 데이터 관계 생성 API", "description": "데이터셋 메타데이터 추출&임베딩&파이프라인(규칙 → 유사도 → LLM 검증)으로 데이터셋 간 관계를 생성하여 Neptune 지식그래프와 OpenSearch에 적재합니다.\n\n- **Datasets**: 데이터셋 등록·조회·삭제\n- **Relations**: 관계 생성 파이프라인·조회·승인 반영\n- **Search**: 유사/자연어/연계 검색과 그래프 시각화\n- **Catalog**: 통합 카탈로그 연동(등록·동기화)\n- **KeyValue Graph**: 레코드 수준 Key-Value 그래프(Level 2)", "license": {"name": "Wisenut"}, "version": "Unknown"}, "paths": {"/keyvalue/batch_extract": {"post": {"tags": ["KeyValue Graph"], "summary": "레코드 배치 Key-Value 추출 및 그래프 생성", "description": "JSON 배열 파일에서 Key-Value를 배치 추출하고 그래프 생성.\n\nArgs:\n    file: JSON 배열 파일 [{...}, {...}, ...]\n    centrality_name: 중심 노드 이름 (예: \"병원현황\", \"판례정보\"). auto_detect_schema=True면 무시됨\n    auto_detect_schema: 스키마 자동 감지 여부 (True: 데이터 분석 후 자동 감지, False: 기본 매핑 사용)\n    sample_size: 스키마 분석 시 샘플링할 레코드 개수 (auto_detect_schema=True일 때만 사용)\n\nReturns:\n    배치 처리 결과 (성공/실패 통계)", "operationId": "batch_extract_key_value_from_json_keyvalue_batch_extract_post", "security": [{"APIKeyHeader": []}], "parameters": [{"name": "centrality_name", "in": "query", "required": false, "schema": {"type": "string", "default": "병원현황", "title": "Centrality Name"}}, {"name": "auto_detect_schema", "in": "query", "required": false, "schema": {"type": "boolean", "default": false, "title": "Auto Detect Schema"}}, {"name": "sample_size", "in": "query", "required": false, "schema": {"type": "integer", "default": 5, "title": "Sample Size"}}, {"name": "table_name", "in": "query", "required": false, "schema": {"type": "string", "default": "", "title": "Table Name"}}], "requestBody": {"required": true, "content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_batch_extract_key_value_from_json_keyvalue_batch_extract_post"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/APIResponseModel"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/keyvalue/normalize": {"post": {"tags": ["KeyValue Graph"], "summary": "클러스터링 기반 Value 자동 정규화", "description": "클러스터링 기반 Value 자동 정규화.\n\nArgs:\n    key_name: 특정 Key명만 정규화 (None이면 전체 Key 정규화)\n    similarity_threshold: 유사도 임계값 (0~1, 높을수록 엄격). 기본값 0.85\n    min_cluster_size: 최소 클러스터 크기 (작은 클러스터 무시). 기본값 2\n\nFlow:\n    1. Key별로 연결된 모든 Value 조회\n    2. Value 간 유사도 행렬 계산 (RapidFuzz token_sort_ratio)\n    3. DBSCAN 클러스터링으로 유사한 Value 그룹화\n    4. 각 클러스터의 대표값 선정 (빈도수 높은 것, 동률이면 긴 것)\n    5. NormalizedValue 노드 생성 및 SAME_AS 관계 생성\n\nReturns:\n    정규화 통계 (Key 수, 클러스터 수, 정규화된 Value 수)", "operationId": "normalize_key_values_keyvalue_normalize_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Body_normalize_key_values_keyvalue_normalize_post"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/APIResponseModel"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"APIKeyHeader": []}]}}, "/keyvalue/analyze_schema": {"post": {"tags": ["KeyValue Graph"], "summary": "데이터 스키마 자동 분석 (LLM)", "description": "데이터 스키마 자동 분석\n\n파일을 업로드하면 샘플링 + 통계 분석을 통해 스키마를 자동으로 추론합니다.\n\nArgs:\n    file: JSON 배열 파일 [{...}, {...}, ...]\n    sample_size: 샘플링할 레코드 개수 (기본: 5)\n\nFlow:\n    1. 파일에서 레코드 추출\n    2. 샘플링 (균등 분포)\n    3. 필드별 통계 수집 (unique 개수, null 비율, 샘플 값)\n    4. LLM으로 스키마 추론 (도메인, 필드 한글명, 데이터 타입)", "operationId": "analyze_data_schema_keyvalue_analyze_schema_post", "requestBody": {"content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_analyze_data_schema_keyvalue_analyze_schema_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/APIResponseModel"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"APIKeyHeader": []}]}}, "/keyvalue/discover_relation_candidates": {"post": {"tags": ["KeyValue Graph"], "summary": "레코드 간 관계 후보 발굴 (Level 2, 온디맨드)", "description": "단일 Document ID로 관계 후보 발굴 (온디맨드 방식)\n\n프로세스:\n1. Document의 Key-Value 메타데이터 분석\n2. LLM이 핵심 Key 선정 및 검색 전략 수립\n3. 그래프 구조 + 벡터 유사도로 후보 검색\n4. 상위 N개 후보 반환 (사용자 검토용)\n\nArgs:\n    source_doc_id: 기준 Document ID\n    top_k: 반환할 후보 개수 (기본 10개)", "operationId": "discover_relation_candidates_keyvalue_discover_relation_candidates_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Body_discover_relation_candidates_keyvalue_discover_relation_candidates_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/APIResponseModel"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"APIKeyHeader": []}]}}, "/workflow/execute": {"post": {"tags": ["KeyValue Graph"], "summary": "파일 업로드 → KV 그래프 워크플로우 실행 (SSE)", "description": "파일 업로드 → 자동 워크플로우 실행 (SSE 스트리밍).\n\nArgs:\n    file: JSON 배열 파일\n    auto_detect_schema: 스키마 자동 감지 여부\n    sample_size: 스키마 분석 샘플 크기\n    similarity_threshold: 정규화 유사도 임계값\n    min_cluster_size: 정규화 최소 클러스터 크기\n    table_name: 데이터셋(테이블)명. Level 1 관계/게이팅과 정합되도록 지정 권장.\n        비우면 업로드 파일명 stem 을 사용.\n\nReturns:\n    Server-Sent Events 스트림", "operationId": "execute_workflow_workflow_execute_post", "requestBody": {"content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_execute_workflow_workflow_execute_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"APIKeyHeader": []}]}}, "/health": {"get": {"tags": ["System"], "summary": "Health", "operationId": "health_health_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/info": {"get": {"tags": ["System"], "summary": "Info", "operationId": "info_info_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}}, "components": {"schemas": {"APIResponseModel": {"properties": {"code": {"type": "integer", "title": "Code", "default": 100200}, "message": {"type": "string", "title": "Message", "default": "API Response Success"}, "result": {"title": "Result", "default": {}}, "description": {"type": "string", "title": "Description", "default": "응답과 관련된 자세한 설명 작성"}}, "type": "object", "title": "APIResponseModel", "description": "기본 API 응답 포맷 by AIP Restful API 디자인 가이드\n\nAI플랫폼팀 API 정식 포맷으로 그대로 사용 권장"}, "Body_analyze_data_schema_keyvalue_analyze_schema_post": {"properties": {"file": {"type": "string", "format": "binary", "title": "File"}, "sample_size": {"type": "integer", "title": "Sample Size", "default": 5}}, "type": "object", "required": ["file"], "title": "Body_analyze_data_schema_keyvalue_analyze_schema_post"}, "Body_batch_extract_key_value_from_json_keyvalue_batch_extract_post": {"properties": {"file": {"type": "string", "format": "binary", "title": "File"}}, "type": "object", "required": ["file"], "title": "Body_batch_extract_key_value_from_json_keyvalue_batch_extract_post"}, "Body_discover_relation_candidates_keyvalue_discover_relation_candidates_post": {"properties": {"source_doc_id": {"type": "string", "title": "Source Doc Id", "description": "기준 Document ID"}, "top_k": {"type": "integer", "title": "Top K", "description": "반환할 후보 개수", "default": 10}}, "type": "object", "required": ["source_doc_id"], "title": "Body_discover_relation_candidates_keyvalue_discover_relation_candidates_post"}, "Body_execute_workflow_workflow_execute_post": {"properties": {"file": {"type": "string", "format": "binary", "title": "File"}, "auto_detect_schema": {"type": "boolean", "title": "Auto Detect Schema", "default": true}, "sample_size": {"type": "integer", "title": "Sample Size", "default": 5}, "similarity_threshold": {"type": "number", "title": "Similarity Threshold", "default": 0.85}, "min_cluster_size": {"type": "integer", "title": "Min Cluster Size", "default": 2}, "table_name": {"type": "string", "title": "Table Name", "default": ""}}, "type": "object", "required": ["file"], "title": "Body_execute_workflow_workflow_execute_post"}, "Body_normalize_key_values_keyvalue_normalize_post": {"properties": {"key_name": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Key Name"}, "similarity_threshold": {"type": "number", "title": "Similarity Threshold", "default": 0.85}, "min_cluster_size": {"type": "integer", "title": "Min Cluster Size", "default": 2}}, "type": "object", "title": "Body_normalize_key_values_keyvalue_normalize_post"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}, "securitySchemes": {"APIKeyHeader": {"type": "apiKey", "in": "header", "name": "x-token"}}}, "tags": [{"name": "Datasets", "description": "데이터셋 등록·조회·삭제·현황"}, {"name": "Search", "description": "유사 테이블 검색, 자연어 검색, 연계 데이터셋 탐색, 부분 그래프 조회"}, {"name": "Relations", "description": "데이터셋 간 관계 생성(3-Tier 파이프라인)·저장·조회"}, {"name": "Catalog", "description": "① 통합 카탈로그 연동 — 엔트리 등록, 변경 동기화, 승인 결정 콜백"}, {"name": "KeyValue Graph", "description": "레코드 수준 Key-Value 그래프 생성·정규화·관계 후보 발굴 (Level 2)"}, {"name": "System", "description": "헬스체크·버전 정보"}]}